pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "textual>=8.2.1",
]
